+++ src/azure-cli-core/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 2.0.3 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 *core: Allow file path of accessTokens.json to be configurable through an env var(#2605)
-
+*core: Allow configured defaults to apply on optional args(#2703)
 
 2.0.2 (2017-04-03)
 ^^^^^^^^^^^^^^^^^^