+++ src/azure-cli-core/HISTORY.rst
@@ -8,6 +8,9 @@ Release History
 *core: Allow file path of accessTokens.json to be configurable through an env var(#2605)
 *core: Allow configured defaults to apply on optional args(#2703)
 *core: Improved performance
+*core: Support for multiple API versions
+*core: Custom CA Certs - Support setting REQUESTS_CA_BUNDLE environment variable
+*core: Cloud configuration - use 'resource manager' endpoint if 'management' endpoint not set
 
 2.0.2 (2017-04-03)
 ^^^^^^^^^^^^^^^^^^